WIP: Current state of parameterization of upper GI reconstruction approach
Comments: The amount of possibilities is striking and makes it difficult to find "best" solution. It seems to be worth setting 25 iterations for the SRR algorithm (here it was 20) -- the rastering effect becomes more apparent but also the boundaries become sharper. Also it might be worth setting the ANTSradius to 30 for the slice-to-volume registration part

Former-commit-id: 790f5a9a4880792c1ac4be34ba871c6e0bf1b66f